[REFACTOR] Change components' name
