--- src/styles/Root.module.css
@@ -1,41 +1,0 @@
-/* Root.css */
-
-#sidebar {
-  width: 200px;
-  height: 100vh; /* 화면 높이에 맞게 사이드바 높이 설정 */
-  position: fixed;
-  top: 0;
-  left: 0;
-  background-color: #f0f0f0;
-  padding: 0px;
-}
-
-#sidebar h1 {
-  font-size: 24px;
-}
-
-#sidebar ul {
-  list-style: none;
-  margin: 0;
-  padding: 0;
-}
-
-#sidebar li {
-  padding: 10px 0;
-}
-
-#sidebar a {
-  text-decoration: none;
-  color: #333;
-  font-weight: bold;
-  display: block;
-}
-
-#detail {
-  margin-left: 250px; /* 사이드바 너비만큼 컨텐츠를 이동시킴 */
-  padding: 20px;
-}
-
-#userProfile {
-  background-color: #ffdd6f;
-}
