--- workflows/main.yml
@@ -1,33 +1,0 @@
-name: peeptodo react CI
-
-on:
-  # Triggers the workflow on push or pull request events but only for the master branch
-  push:
-    branches: [ develop ]
-    
-  # Allows you to run this workflow manually from the Actions tab
-  workflow_dispatch:
-  
-# A workflow run is made up of one or more jobs that can run sequentially or in parallel
-jobs:
-  # This workflow contains a single job called "build"
-  build:
-    # The type of runner that the job will run on
-    runs-on: ubuntu-latest
-    # Steps represent a sequence of tasks that will be executed as part of the job
-    
-    steps:
-      # Checks-out your repository under $GITHUB_WORKSPACE, so your job can access it
-      
-      - uses: actions/checkout@v3
-      
-      - name: Auth to GCP
-        uses: 'google-github-actions/auth@v1'
-        with:
-          credentials_json: ${{ secrets.GCP_SA_KEY }}
-          
-      - name: Deploy to App Engine
-        uses: 'google-github-actions/deploy-appengine@v1'
-        with:
-          deliverables: app.yaml
-          project_id: ${{ secrets.PROJECT_ID }}
